# Conflicts:
#	go.mod
#	go.sum

+++ cmd/mirror-main.go
@@ -145,6 +145,10 @@ var (
 			Name:  "skip-errors",
 			Usage: "skip any errors when mirroring",
 		},
+		cli.IntFlag{
+			Name:  "max-workers",
+			Usage: "maximum number of concurrent copies (default: autodetect)",
+		},
 		checksumFlag,
 	}
 )
@@ -745,7 +749,7 @@ func (mj *mirrorJob) watchMirrorEvents(ctx context.Context, events []EventInfo) 
 			}
 			mirrorURL.TotalCount = mj.status.GetCounts()
 			mirrorURL.TotalSize = mj.status.Get()
-			if mirrorURL.TargetContent != nil && (mj.opts.isRemove || mj.opts.activeActive) {
+			if mirrorURL.TargetContent != nil && (mj.opts.isRemove || mj.opts.activeActive || mj.opts.isWatch) {
 				mj.parallel.queueTask(func() URLs {
 					return mj.doRemove(ctx, mirrorURL, event)
 				}, 0)
@@ -867,8 +871,10 @@ func (mj *mirrorJob) mirror(ctx context.Context) bool {
 	var wg sync.WaitGroup
 	ctx, cancel := context.WithCancel(ctx)
 
+	doneCh := make(chan struct{})
+
 	// Starts watcher loop for watching for new events.
-	if mj.opts.isWatch {
+	if mj.opts.isWatch || mj.opts.activeActive {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
@@ -889,9 +895,12 @@ func (mj *mirrorJob) mirror(ctx context.Context) bool {
 		wg.Wait()
 		mj.parallel.stopAndWait()
 		close(mj.statusCh)
+		close(doneCh)
 	}()
 
-	return mj.monitorMirrorStatus(cancel)
+	ret := mj.monitorMirrorStatus(cancel)
+	<-doneCh
+	return ret
 }
 
 func newMirrorJob(srcURL, dstURL string, opts mirrorOptions) *mirrorJob {
@@ -905,7 +914,7 @@ func newMirrorJob(srcURL, dstURL string, opts mirrorOptions) *mirrorJob {
 		watcher:   NewWatcher(UTCNow()),
 	}
 
-	mj.parallel = newParallelManager(mj.statusCh)
+	mj.parallel = newParallelManager(mj.statusCh, opts.maxWorkers)
 
 	// we'll define the status to use here,
 	// do we want the quiet status? or the progressbar
@@ -993,7 +1002,8 @@ func runMirror(ctx context.Context, srcURL, dstURL string, cli *cli.Context, enc
 		isOverwrite = cli.Bool("overwrite")
 	}
 
-	isWatch := cli.Bool("watch") || cli.Bool("multi-master") || cli.Bool("active-active")
+	isWatch := cli.Bool("watch") || cli.Bool("multi-master")
+	isActiveActive := cli.Bool("active-active")
 	isRemove := cli.Bool("remove")
 	md5, checksum := parseChecksum(cli)
 
@@ -1021,7 +1031,8 @@ func runMirror(ctx context.Context, srcURL, dstURL string, cli *cli.Context, enc
 		storageClass:          cli.String("storage-class"),
 		userMetadata:          userMetadata,
 		encKeyDB:              encKeyDB,
-		activeActive:          isWatch,
+		activeActive:          isActiveActive,
+		maxWorkers:            cli.Int("max-workers"),
 	}
 
 	// If we are not using active/active and we are not removing
@@ -1053,7 +1064,7 @@ func runMirror(ctx context.Context, srcURL, dstURL string, cli *cli.Context, enc
 
 			if d.Diff == differInSecond {
 				diffBucket := strings.TrimPrefix(d.SecondURL, dstClt.GetURL().String())
-				if !isFake && isRemove {
+				if !isFake && isRemove && createDstBuckets {
 					aliasedDstBucket := path.Join(dstURL, diffBucket)
 					err := deleteBucket(ctx, aliasedDstBucket, false)
 					mj.status.fatalIf(err, "Failed to start mirroring.")
@@ -1070,6 +1081,15 @@ func runMirror(ctx context.Context, srcURL, dstURL string, cli *cli.Context, enc
 			newDstClt, _ := newClient(newTgtURL)
 
 			if d.Diff == differInFirst {
+				// This loop is responsible solely for bringing target on par with source at the bucket level.
+				// createDstBuckets == false implies that the target itself represents a bucket.
+				// So we don't want to perform MakeBucket (and in turn delete and recreate already existing buckets and objects) if it already exists.
+				// Any differences at the object level between source and target's bucket will be synced as part of objectDifference() that gets
+				// called later.
+				if !createDstBuckets {
+					continue
+				}
+
 				var (
 					withLock bool
 					mode     minio.RetentionMode
@@ -1123,7 +1143,7 @@ func runMirror(ctx context.Context, srcURL, dstURL string, cli *cli.Context, enc
 		}
 	}
 
-	if mj.opts.isWatch {
+	if mj.opts.isWatch || mj.opts.activeActive {
 		// monitor mode will watch the source folders for changes,
 		// and queue them for copying.
 		if err := mj.watchURL(ctx, srcClt); err != nil {
